[Step3 - 자동차 경주] refactor: interface 제거
