+++ src/main/java/step3/view/ResultView.java
@@ -1,9 +1,21 @@
 package step3.view;
 
-import step3.Car;
+import step3.Cars;
 
-import java.util.List;
+import java.util.stream.Collectors;
 
-public interface ResultView {
-  void print(List<Car> cars);
+public class ResultView {
+
+  private ResultView() {};
+
+  public void print (Cars cars) {
+    System.out.println(
+      cars.stream()
+          .map(car -> car.getPositionString())
+          .collect(Collectors.joining("\n"))
+    );
+  }
+
+  public static ResultView of () { return new ResultView(); }
+
 }
